Move generic component to more generic location
